% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_function.R
\name{add_a_and_b}
\alias{add_a_and_b}
\title{Add Two Numbers}
\usage{
add_a_and_b(a, b)
}
\arguments{
\item{a}{First number}

\item{b}{Second number}
}
\value{
The sum of \code{a} and \code{b}.
}
\description{
This function returns the sum of two numbers.
}
\examples{
add_a_and_b(2, 3)
add_a_and_b(10, -5)
}
