% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{downloadReportHandler}
\alias{downloadReportHandler}
\title{This function of the Shiny server builds a PDF report that summarizes the
scores from the YSQ, SMI, and SCI questionnaires. It uses an existing R
Markdown file (`Intake Schema Therapy Report.Rmd`) as the template.}
\usage{
downloadReportHandler(get_scores)
}
\arguments{
\item{get_scores}{A function that returns a list with `ysq`, `smi`, and `sci`
scores.}
}
\value{
A Shiny `downloadHandler` object that allows users to download the
report as a PDF file.
}
\description{
The report includes:
- The scores on each schema, schema mode and coping style.
- Descriptions of schemas, modes, and coping styles
}
\examples{
output$downloadReport <- downloadReportHandler(get_scores)

}
