% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_responses.R
\name{downloadResponsesHandler}
\alias{downloadResponsesHandler}
\title{This function of the Shiny server generates a download handler that writes
all responses from YSQ, SMI, and SCI questionnaires into a CSV file.
It includes the questionnaire type, schema/mode/coping category, question
text, response, and the average score for that category.}
\usage{
downloadResponsesHandler(ysq_questions, smi_questions, sci_questions, input)
}
\arguments{
\item{ysq_questions}{YSQ questions}

\item{smi_questions}{SMI questions}

\item{sci_questions}{SCI questions}

\item{input}{Shiny input object containing user responses.}
}
\description{
This function of the Shiny server generates a download handler that writes
all responses from YSQ, SMI, and SCI questionnaires into a CSV file.
It includes the questionnaire type, schema/mode/coping category, question
text, response, and the average score for that category.
}
\examples{
output$downloadResponses <- downloadResponsesHandler(ysq_questions, smi_questions, sci_questions, input)

}
