% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_function.R
\name{generate_noise}
\alias{generate_noise}
\title{Generate random numbers}
\usage{
generate_noise(n)
}
\arguments{
\item{n}{Number of random values to generate}
}
\value{
A numeric vector of random values
}
\description{
Generate random numbers
}
\examples{
generate_noise(5)

}
