% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_handler.R
\name{navHandler}
\alias{navHandler}
\title{This function watches for a button click to move between pages of questions.
It makes sure all questions on the current page are answered before allowing
the user to go to the next page or tab. If questions are missing, it shows
a popup asking the user to complete them.}
\usage{
navHandler(input_id, questions, page_val, next_tab, input_prefix, key)
}
\arguments{
\item{input_id}{The ID of the navigation button to listen for (e.g., "next_ysq").}

\item{questions}{Returns the list of questions for the section.}

\item{page_val}{Stores the current page number.}

\item{next_tab}{The ID of the tab to switch to after finishing all pages.}

\item{input_prefix}{A short label used at the start of input IDs to group
related questions (like "ysq" or "smi").}

\item{key}{The name of the question field that helps build each input’s
unique ID (for example, "schema" or "mode").}
}
\description{
This function watches for a button click to move between pages of questions.
It makes sure all questions on the current page are answered before allowing
the user to go to the next page or tab. If questions are missing, it shows
a popup asking the user to complete them.
}
\examples{
\dontrun{
navHandler(
  input_id = "next_ysq",
  questions = ysq_questions,
  page_val = ysq_page,
  next_tab = "smi",
  input_prefix = "ysq",
  key = "schema"
)
}

}
