% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_question_ui.R
\name{renderQuestionUI}
\alias{renderQuestionUI}
\title{This function creates UI elements to show a set of questions for the current
page. Each question is displayed with radio buttons for answers from 1 to 7.}
\usage{
renderQuestionUI(questions, type, page, key)
}
\arguments{
\item{questions}{Returns a list of question objects.}

\item{type}{A short word added at the start of each input’s ID to keep them
unique, like "ysq" or "smi".}

\item{page}{Indicates the current page number.}

\item{key}{The name of the part in each question used to make unique input
IDs, like 'schema' or 'mode'.}
}
\value{
A list of shiny UI elements showing the questions and answers.
}
\description{
This function creates UI elements to show a set of questions for the current
page. Each question is displayed with radio buttons for answers from 1 to 7.
}
\examples{
if (interactive()) {
  library(shiny)
  questions <- reactive({
    list(
      list(question = "How often do you feel stressed?", schema = "stress", index = 1),
      list(question = "Do you enjoy social events?", schema = "social", index = 2)
    )
  })
  page <- reactiveVal(1)
  ui <- fluidPage(
    uiOutput("questions_ui")
  )
  server <- function(input, output, session) {
    output$questions_ui <- renderUI({
      renderQuestionUI(questions, "ysq", page, "schema")
    })
  }
  shinyApp(ui, server)
}

}
