% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tab.R
\name{report_tab_ui}
\alias{report_tab_ui}
\title{The Report Download tab in the UI builds the UI elements for the Download
Report tab in the Schema Therapy Assessment Shiny app.}
\usage{
report_tab_ui()
}
\value{
A UI object containing:
\itemize{
  \item Title "Download Your Schema Report"
  \item A download button for the .pdf report ("downloadReport")
  \item Line breaks for spacing
  \item A download button for the .csv file ("downloadResponses")
}
}
\description{
The Report Download tab in the UI builds the UI elements for the Download
Report tab in the Schema Therapy Assessment Shiny app.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinydashboard)
  shinyApp(
    ui = dashboardPage(
      dashboardHeader(title = "Test Report"),
      dashboardSidebar(),
      dashboardBody(
        report_tab_ui()
      )
    ),
    server = function(input, output, session) {}
  )
}

}
