% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_complete.R
\name{server_complete}
\alias{server_complete}
\title{The server function :
- Initializes and shuffles questions from YSQ, SMI, and SCI.
- Shows a page of questions with answer options and handles switching between
pages.
- Checks that all questions on the page are answered before moving to the
next section.
- Calculates average scores for each schema, mode, and coping style.
- Allows to download the report in .pdf and the responses in .csv.}
\usage{
server_complete(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
The server function :
- Initializes and shuffles questions from YSQ, SMI, and SCI.
- Shows a page of questions with answer options and handles switching between
pages.
- Checks that all questions on the page are answered before moving to the
next section.
- Calculates average scores for each schema, mode, and coping style.
- Allows to download the report in .pdf and the responses in .csv.
}
\details{
This function expects the following global variables to be available:
- ysq_items, smi_items, sci_items: Lists of questions
- schemas_descriptions, modes_descriptions, coping_descriptions: Description
text for each category.

It creates internal reactive values for:
- Shuffled questions: ysq_questions, smi_questions, sci_questions.
- Page numbers for navigation: ysq_page, smi_page, sci_page.

Also defines helper functions to:
- Show questions on the screen,
- Check that all questions are answered before moving to the next section.
}
