% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smi_tab.R
\name{smi_tab_ui}
\alias{smi_tab_ui}
\title{The schema Mode Inventory tab of the UI builds the UI elements for the Schema
Mode Inventory tab in the Schema Therapy Assessment Shiny app.}
\usage{
smi_tab_ui()
}
\value{
A UI element with:
- A title "Schema Mode Inventory"
- A UI output placeholder "smi_questions_ui" for questions
- Previous and Next navigation buttons
- A text output "smi_page_info" to show the current page number
}
\description{
The schema Mode Inventory tab of the UI builds the UI elements for the Schema
Mode Inventory tab in the Schema Therapy Assessment Shiny app.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinydashboard)
  shinyApp(
    ui = dashboardPage(
      dashboardHeader(title = "Test SMI"),
      dashboardSidebar(),
      dashboardBody(
        smi_tab_ui()
      )
    ),
    server = function(input, output, session) {}
  )
}

}
