% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_complete.R
\name{ui_complete}
\alias{ui_complete}
\title{The complete UI for my Schema Therapy Assessment App builds the full dashboard
layout for the Schema Therapy Assessment Shiny app.}
\usage{
ui_complete()
}
\value{
A shinydashboard containing the complete UI.
}
\description{
The complete UI for my Schema Therapy Assessment App builds the full dashboard
layout for the Schema Therapy Assessment Shiny app.
}
\details{
The layout includes:
\itemize{
  \item A red dashboard
  \item A sidebar menu for navigating between questionnaires and the download of the report
  \item The tab content for the YSQ, SMI, SCI, and report download
}
}
\examples{
if (interactive()) {
  shinyApp(
    ui = ui_complete(),
    server = function(input, output, session) {}
  )
}

}
