% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ysq_tab.R
\name{ysq_tab_ui}
\alias{ysq_tab_ui}
\title{The Young Schema Questionnaire tab of the UI builds the UI elements for the
Young Schema Questionnaire tab in the Schema Therapy Assessment Shiny app.}
\usage{
ysq_tab_ui()
}
\value{
A UI element with:
- A title "Young Schema Questionnaire"
- A UI output placeholder "ysq_questions_ui" for the questions
- Previous and Next buttons for navigation
- A text output "ysq_page_info" to show the current page number
}
\description{
The Young Schema Questionnaire tab of the UI builds the UI elements for the
Young Schema Questionnaire tab in the Schema Therapy Assessment Shiny app.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinydashboard)
  shinyApp(
    ui = dashboardPage(
      dashboardHeader(title = "Test YSQ"),
      dashboardSidebar(),
      dashboardBody(
        ysq_tab_ui()
      )
    ),
    server = function(input, output, session) {}
  )
}

}
